flash_writer_tool: Allow script to be run from anywhere
At the moment if using the flash_writer_tool.sh script from a different
location from where the script file is located there can be issues as when
the script calls itself it is hardcoded to ./flash_writer_tool.sh.

Fix this by working out the script location and use the absolute path to
call the flash_writer_tool.sh script.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>