build_scripts: Remove unused defconfig stuff
The make_config command was something left over from the first RZ/G2 BSP.
It's out of scope now a days.

Also, there is no reason to define DEFCONFIG="xxx", because all the devices
now use the same file.

Also, downloading the firmware files for USB3, Wifi and Bluetooth are out
of scope for this script.